requirements.txt
python-socketio==1.8.0
Flask==0.12.2
eventlet==0.21.0
redis==2.10.6
